# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EDU Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /certificates/{serial}:
    get:
      operationId: getCertificate
      tags: [certificates]
      summary: Verify a certificate by serial (public)
      parameters:
        - { name: serial, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Certificate
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Certificate" }

  # ── LMS: REVIEWS ─────────────────────────────────────────────────
  /courses/{courseId}/reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews and rating summary for a course
      parameters:
        - { name: courseId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Reviews summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewSummary" }
    post:
      operationId: upsertReview
      tags: [reviews]
      summary: Create or update the current user review (enrolled only)
      parameters:
        - { name: courseId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReviewInput" }
      responses:
        "200":
          description: Saved review
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Review" }

  # ── LMS: NOTES ───────────────────────────────────────────────────
  /lessons/{lessonId}/notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List current user notes for a lesson
      parameters:
        - { name: lessonId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Note" }
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note on a lesson
      parameters:
        - { name: lessonId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteInput" }
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Note" }
  /notes/{noteId}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - { name: noteId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteUpdate" }
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Note" }
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - { name: noteId, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }

  # ── LMS: BOOKMARKS ───────────────────────────────────────────────
  /courses/{courseId}/bookmarks:
    get:
      operationId: listBookmarks
      tags: [bookmarks]
      summary: List current user bookmarks for a course
      parameters:
        - { name: courseId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Bookmarks list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Bookmark" }
  /lessons/{lessonId}/bookmark:
    post:
      operationId: toggleBookmark
      tags: [bookmarks]
      summary: Toggle a bookmark on a lesson
      parameters:
        - { name: lessonId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Toggle result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookmarkToggleResult" }

  # ── LMS: PREREQUISITES ───────────────────────────────────────────
...
components:
  schemas:
...
    Certificate:
      type: object
      required: [serial, courseId, userId, issuedAt]
      properties:
        serial: { type: string }
        courseId: { type: integer }
        userId: { type: string }
        issuedAt: { type: string }
        courseTitle: { type: ["string", "null"] }
        userName: { type: ["string", "null"] }
        instructorName: { type: ["string", "null"] }

    # ── LMS: REVIEWS ───────────────────────────────────────────────
    Review:
      type: object
      required: [id, courseId, userId, rating, createdAt]
      properties:
        id: { type: integer }
        courseId: { type: integer }
        userId: { type: string }
        userName: { type: ["string", "null"] }
        rating: { type: integer }
        comment: { type: ["string", "null"] }
        createdAt: { type: string }
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
    ReviewSummary:
      type: object
      required: [average, count]
      properties:
        average: { type: number }
        count: { type: integer }
        distribution:
          type: array
          items: { type: integer }
        myReview: { $ref: "#/components/schemas/Review" }
        reviews:
          type: array
          items: { $ref: "#/components/schemas/Review" }

    # ── LMS: NOTES & BOOKMARKS ─────────────────────────────────────
    Note:
      type: object
      required: [id, lessonId, courseId, userId, content, createdAt]
      properties:
        id: { type: integer }
        lessonId: { type: integer }
        courseId: { type: integer }
        userId: { type: string }
        content: { type: string }
        timestampSeconds: { type: ["integer", "null"] }
        createdAt: { type: string }
    NoteInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
        timestampSeconds: { type: integer }
    NoteUpdate:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
    Bookmark:
      type: object
      required: [lessonId, courseId, userId]
      properties:
        id: { type: integer }
        lessonId: { type: integer }
        courseId: { type: integer }
        userId: { type: string }
        createdAt: { type: string }
        lessonTitle: { type: ["string", "null"] }
    BookmarkToggleResult:
      type: object
      required: [lessonId, bookmarked]
      properties:
        lessonId: { type: integer }
        bookmarked: { type: boolean }

    # ── LMS: PREREQUISITES ─────────────────────────────────────────
...